pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "danish_to_english_llm"
version = "0.0.1"
description = "Translate short sentences from Danish to English using LLM"
authors = [
  { name = "Group 26", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11"

# Remove `dependencies` from `dynamic` list
dynamic = ["optional-dependencies"]
dependencies = [
    "datasets>=3.2.0",
    "docker>=7.1.0",
    "mypy>=1.14.1",
    "pytorch-lightning>=2.5.0.post0",
    "ruff>=0.1.3",
    "sentencepiece>=0.2.0",
    "setuptools>=75.8.0",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "transformers>=4.47.1",
    "wandb>=0.19.2",
]

[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}

[tool.setuptools.dynamic.optional-dependencies]
dev = {file = ['requirements_dev.txt']}

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
